feat: много маленьких правок при подготовке к развёртыванию

+++ sso/src/main/java/vlad/erofeev/sso/domain/Person.java
@@ -22,7 +22,7 @@ public class Person {
 
     private String email;
     private String password;
-    private Integer phone;
+    private Long phone;
     private String addr;
     private String name;
     private String surname;
